changed app.js and optimized code

+++ public/theme/build/js/app.js
@@ -1,6 +1,9 @@
 console.log('yes app.js');
-let csrf = document.querySelector('meta[name="csrf_token"]').content;
-let tabId = document.querySelector('table').id;
+
+const msgTimeOut = 500;
+const msgFadeOut = 7000;
+const csrf = document.querySelector('meta[name="csrf_token"]').content;
+
 $('#master-check').on('click', function (e) {
     if ($(this).is(':checked', true)) {
         $(".sub_chk").prop('checked', true);
@@ -23,6 +26,7 @@ deleteRecord = (moduleName, delUrl) => {
 }
 
 deleteSingle = (dataObj) => {
+    let tabId = document.querySelector('table').id;
     $('#myActionModal').modal('hide');
     $.ajax({
         url: dataObj.currentURl,
@@ -34,14 +38,15 @@ deleteSingle = (dataObj) => {
             $(`#${tabId}`).DataTable().ajax.reload();
             $("#master-check").prop('checked', false);
             if (responseData['responseStatus'] === 1) {
-                $("#message-green").show().fadeOut(7000);
-                document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} deleted successfully!`;
-            } else {
-                $("#message-red").show().fadeOut(7000);
-                document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
                 setTimeout(() => {
-                    location.reload();
-                }, 3000);
+                    $("#message-green").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} deleted successfully!`;
+                }, msgTimeOut);
+            } else {
+                setTimeout(() => {
+                    $("#message-red").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
+                }, msgTimeOut);
             }
         },
         error: function (responseData) {
@@ -51,51 +56,55 @@ deleteSingle = (dataObj) => {
 }
 
 // Active Action
-document.querySelector('.active_all').addEventListener("click", (e) => {
-    let allVals = [];
-    let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
+let activeAll = document.querySelector('.active_all');
+if (activeAll) {
+    activeAll.addEventListener("click", (e) => {
+        let allVals = [];
+        let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
 
-    checkboxes.forEach(element => {
-        let ids = element.getAttribute('data-id');
-        allVals.push(ids);
+        checkboxes.forEach(element => {
+            let ids = element.getAttribute('data-id');
+            allVals.push(ids);
+        });
+
+        if (allVals.length > 1) {
+            if (e.target.getAttribute('data-pluraltext')) {
+                var moduleName = e.target.getAttribute('data-pluraltext');
+            } else {
+                var moduleName = 'rows';
+            }
+            var values = 'these';
+        } else {
+            if (e.target.getAttribute('data-text')) {
+                var moduleName = e.target.getAttribute('data-text');
+            } else {
+                var moduleName = 'row';
+            }
+            var values = 'this';
+        }
+
+        if (allVals.length <= 0) {
+            $("#message-red").show().fadeOut(msgFadeOut);
+            document.getElementById('message-red-err').innerHTML = 'Error! Please select record to make a status Active.';
+        }
+        else {
+            let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
+            let stringObj = JSON.stringify(dataObj);
+            $('#myActionModal').modal('show');
+            $("#myActionModalLabel").text('Active');
+            document.getElementById('myActionModalContent').innerText = `Active`;
+            document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to active ${values} ${moduleName} ?`;
+            document.getElementById('MyActionModalBtn').setAttribute('onclick', `makeActive("${allVals}",${stringObj})`);
+            document.getElementById("MyActionModalBtn").className = "btn btn-success";
+            document.getElementById('MyActionModalBtn').innerHTML = `Make Active`;
+        }
+
     });
-
-    if (allVals.length > 1) {
-        if (e.target.getAttribute('data-pluraltext')) {
-            var moduleName = e.target.getAttribute('data-pluraltext');
-        } else {
-            var moduleName = 'rows';
-        }
-        var values = 'these';
-    } else {
-        if (e.target.getAttribute('data-text')) {
-            var moduleName = e.target.getAttribute('data-text');
-        } else {
-            var moduleName = 'row';
-        }
-        var values = 'this';
-    }
-
-    if (allVals.length <= 0) {
-        $("#message-red").show().fadeOut(7000);
-        document.getElementById('message-red-err').innerHTML = 'Error! Please select record to make a status Active.';
-    }
-    else {
-        let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
-        let stringObj = JSON.stringify(dataObj);
-        $('#myActionModal').modal('show');
-        $("#myActionModalLabel").text('Active');
-        document.getElementById('myActionModalContent').innerText = `Active`;
-        document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to active ${values} ${moduleName} ?`;
-        document.getElementById('MyActionModalBtn').setAttribute('onclick', `makeActive("${allVals}",${stringObj})`);
-        document.getElementById("MyActionModalBtn").className = "btn btn-success";
-        document.getElementById('MyActionModalBtn').innerHTML = `Make Active`;
-    }
-
-});
+}
 
 
 makeActive = (allVals, dataObj) => {
+    let tabId = document.querySelector('table').id;
     $('#myActionModal').modal('hide');
     $.ajax({
         url: dataObj.currentURl,
@@ -106,14 +115,15 @@ makeActive = (allVals, dataObj) => {
             $(`#${tabId}`).DataTable().ajax.reload();
             $("#master-check").prop('checked', false);
             if (responseData['responseStatus'] === 1) {
-                $("#message-green").show().fadeOut(7000);
-                document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} activated successfully!`;
-            } else {
-                $("#message-red").show().fadeOut(7000);
-                document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
                 setTimeout(() => {
-                    location.reload();
-                }, 3000);
+                    $("#message-green").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} activated successfully!`;
+                }, msgTimeOut);
+            } else {
+                setTimeout(() => {
+                    $("#message-red").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
+                }, msgTimeOut);
             }
         },
         error: function (responseData) {
@@ -124,51 +134,55 @@ makeActive = (allVals, dataObj) => {
 
 
 // Inactive Action
-document.querySelector('.Inactive_all').addEventListener("click", (e) => {
-    let allVals = [];
-    let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
+let inactiveAll = document.querySelector('.Inactive_all');
+if (inactiveAll) {
+    inactiveAll.addEventListener("click", (e) => {
+        let allVals = [];
+        let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
 
-    checkboxes.forEach(element => {
-        let ids = element.getAttribute('data-id');
-        allVals.push(ids);
+        checkboxes.forEach(element => {
+            let ids = element.getAttribute('data-id');
+            allVals.push(ids);
+        });
+
+        if (allVals.length > 1) {
+            if (e.target.getAttribute('data-pluraltext')) {
+                var moduleName = e.target.getAttribute('data-pluraltext');
+            } else {
+                var moduleName = 'rows';
+            }
+            var values = 'these';
+        } else {
+            if (e.target.getAttribute('data-text')) {
+                var moduleName = e.target.getAttribute('data-text');
+            } else {
+                var moduleName = 'row';
+            }
+            var values = 'this';
+        }
+
+        if (allVals.length <= 0) {
+            $("#message-red").show().fadeOut(msgFadeOut);
+            document.getElementById('message-red-err').innerHTML = 'Error! Please select record to make a status Inactive.';
+        }
+        else {
+            let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
+            let stringObj = JSON.stringify(dataObj);
+            $('#myActionModal').modal('show');
+            $("#myActionModalLabel").text('Inactive');
+            document.getElementById('myActionModalContent').innerText = `Inactive`;
+            document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to Inactive ${values} ${moduleName} ?`;
+            document.getElementById('MyActionModalBtn').setAttribute('onclick', `makeInactive("${allVals}",${stringObj})`);
+            document.getElementById("MyActionModalBtn").className = "btn btn-warning";
+            document.getElementById('MyActionModalBtn').innerHTML = `Make Inactive`;
+        }
+
     });
-
-    if (allVals.length > 1) {
-        if (e.target.getAttribute('data-pluraltext')) {
-            var moduleName = e.target.getAttribute('data-pluraltext');
-        } else {
-            var moduleName = 'rows';
-        }
-        var values = 'these';
-    } else {
-        if (e.target.getAttribute('data-text')) {
-            var moduleName = e.target.getAttribute('data-text');
-        } else {
-            var moduleName = 'row';
-        }
-        var values = 'this';
-    }
-
-    if (allVals.length <= 0) {
-        $("#message-red").show().fadeOut(7000);
-        document.getElementById('message-red-err').innerHTML = 'Error! Please select record to make a status Inactive.';
-    }
-    else {
-        let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
-        let stringObj = JSON.stringify(dataObj);
-        $('#myActionModal').modal('show');
-        $("#myActionModalLabel").text('Inactive');
-        document.getElementById('myActionModalContent').innerText = `Inactive`;
-        document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to Inactive ${values} ${moduleName} ?`;
-        document.getElementById('MyActionModalBtn').setAttribute('onclick', `makeInactive("${allVals}",${stringObj})`);
-        document.getElementById("MyActionModalBtn").className = "btn btn-warning";
-        document.getElementById('MyActionModalBtn').innerHTML = `Make Inactive`;
-    }
-
-});
+}
 
 
 makeInactive = (allVals, dataObj) => {
+    let tabId = document.querySelector('table').id;
     $('#myActionModal').modal('hide');
     $.ajax({
         url: dataObj.currentURl,
@@ -179,14 +193,15 @@ makeInactive = (allVals, dataObj) => {
             $(`#${tabId}`).DataTable().ajax.reload();
             $("#master-check").prop('checked', false);
             if (responseData['responseStatus'] === 1) {
-                $("#message-green").show().fadeOut(7000);
-                document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} inactivated successfully!`;
-            } else {
-                $("#message-red").show().fadeOut(7000);
-                document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
                 setTimeout(() => {
-                    location.reload();
-                }, 3000);
+                    $("#message-green").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} inactivated successfully!`;
+                }, msgTimeOut);
+            } else {
+                setTimeout(() => {
+                    $("#message-red").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
+                }, msgTimeOut);
             }
         },
         error: function (responseData) {
@@ -197,51 +212,55 @@ makeInactive = (allVals, dataObj) => {
 
 
 // DeleteAll Action
-document.querySelector('.delete_all').addEventListener("click", (e) => {
-    let allVals = [];
-    let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
+let deleteAll = document.querySelector('.delete_all');
+if (deleteAll) {
+    deleteAll.addEventListener("click", (e) => {
+        let allVals = [];
+        let checkboxes = document.querySelectorAll('input[name="sub_chk"]:checked');
 
-    checkboxes.forEach(element => {
-        let ids = element.getAttribute('data-id');
-        allVals.push(ids);
+        checkboxes.forEach(element => {
+            let ids = element.getAttribute('data-id');
+            allVals.push(ids);
+        });
+
+        if (allVals.length > 1) {
+            if (e.target.getAttribute('data-pluraltext')) {
+                var moduleName = e.target.getAttribute('data-pluraltext');
+            } else {
+                var moduleName = 'rows';
+            }
+            var values = 'these';
+        } else {
+            if (e.target.getAttribute('data-text')) {
+                var moduleName = e.target.getAttribute('data-text');
+            } else {
+                var moduleName = 'row';
+            }
+            var values = 'this';
+        }
+
+        if (allVals.length <= 0) {
+            $("#message-red").show().fadeOut(msgFadeOut);
+            document.getElementById('message-red-err').innerHTML = 'Error! Please select record to delete.';
+        }
+        else {
+            let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
+            let stringObj = JSON.stringify(dataObj);
+            $('#myActionModal').modal('show');
+            $("#myActionModalLabel").text('Delete');
+            document.getElementById('myActionModalContent').innerText = `Delete`;
+            document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to detele ${values} ${moduleName} ?`;
+            document.getElementById('MyActionModalBtn').setAttribute('onclick', `DeleteAll("${allVals}",${stringObj})`);
+            document.getElementById("MyActionModalBtn").className = "btn btn-danger";
+            document.getElementById('MyActionModalBtn').innerHTML = `Delete`;
+        }
+
     });
-
-    if (allVals.length > 1) {
-        if (e.target.getAttribute('data-pluraltext')) {
-            var moduleName = e.target.getAttribute('data-pluraltext');
-        } else {
-            var moduleName = 'rows';
-        }
-        var values = 'these';
-    } else {
-        if (e.target.getAttribute('data-text')) {
-            var moduleName = e.target.getAttribute('data-text');
-        } else {
-            var moduleName = 'row';
-        }
-        var values = 'this';
-    }
-
-    if (allVals.length <= 0) {
-        $("#message-red").show().fadeOut(7000);
-        document.getElementById('message-red-err').innerHTML = 'Error! Please select record to delete.';
-    }
-    else {
-        let dataObj = { currentURl: e.target.getAttribute('data-url'), moduleName: moduleName };
-        let stringObj = JSON.stringify(dataObj);
-        $('#myActionModal').modal('show');
-        $("#myActionModalLabel").text('Delete');
-        document.getElementById('myActionModalContent').innerText = `Delete`;
-        document.getElementById('myActionModalContent').innerHTML = `Are you sure you want to detele ${values} ${moduleName} ?`;
-        document.getElementById('MyActionModalBtn').setAttribute('onclick', `DeleteAll("${allVals}",${stringObj})`);
-        document.getElementById("MyActionModalBtn").className = "btn btn-danger";
-        document.getElementById('MyActionModalBtn').innerHTML = `Delete`;
-    }
-
-});
+}
 
 
 DeleteAll = (allVals, dataObj) => {
+    let tabId = document.querySelector('table').id;
     $('#myActionModal').modal('hide');
     $.ajax({
         url: dataObj.currentURl,
@@ -252,14 +271,15 @@ DeleteAll = (allVals, dataObj) => {
             $(`#${tabId}`).DataTable().ajax.reload();
             $("#master-check").prop('checked', false);
             if (responseData['responseStatus'] === 1) {
-                $("#message-green").show().fadeOut(7000);
-                document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} deleted successfully!`;
-            } else {
-                $("#message-red").show().fadeOut(7000);
-                document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
                 setTimeout(() => {
-                    location.reload();
-                }, 3000);
+                    $("#message-green").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-green-succ').innerHTML = `Success! ${dataObj.moduleName} deleted successfully!`;
+                }, msgTimeOut);
+            } else {
+                setTimeout(() => {
+                    $("#message-red").show().fadeOut(msgFadeOut);
+                    document.getElementById('message-red-err').innerHTML = 'Error! Oops..some error occured!';
+                }, msgTimeOut);
             }
         },
         error: function (responseData) {
